back/requirements/base.txt
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2-binary==2.9.1
django-jazzmin
Markdown==3.3.4

asgiref==3.3.4
autopep8==1.5.7
defusedxml==0.7.1
diff-match-patch==20200713
Django-Admin-Object-Actions==0.1.11
django-cprofile-middleware==1.0.5
django-crum==0.7.9
django-import-export==2.5.0
django-object-actions==3.0.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pycodestyle==2.7.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
xlrd==2.0.1
xlwt==1.3.0
